+++ docs/CHANGELOG.md
@@ -1,5 +1,24 @@
 # changelog
 
+## Version 1.5.0 (20210610)
+### Added
+* Added silent update check on script start (info onlx if update available)
+* Adding ascii-art to 
+  * youtube-dl update check
+  * download start
+
+### Changes
+* Improved version check to handle all 3 cases (up-to-date VS outdated VS dev version)
+* Download UI: changed button label
+* dialogs all show the project url using the --hline parameter
+* dialogs are now using the same width
+* Improve url input validation (now checks if it is a link & if so if its reachable)
+
+### Removed
+* Removed zenity dependency
+
+
+
 ## Version 1.4.1 (20210530)
 ### Changes
 * Simplify dialogs 
@@ -9,12 +28,14 @@
 * Fix ydownl.sh update script (if version is up to date)
 
 
+
 ## Version 1.4.0 (20210527)
 * Switched to ncruses like mode based on 'dialog'
 * Added variable for download target folder
 * Youtube-DL: Removing --newline
 * Youtube-DL: Adding update check before execution
 * Youtube-DL: Adding --no-mtime
+
 
 
 ## Version 1.3.0 (20210402)